% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{ash}
\alias{ash}
\title{asinh data transformation}
\usage{
ash(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
The asinh trafo, an alternative to log transformation that has less
(compressing) effects on the extreme values (low and high values),
and naturally handles negative numbers and 0.
}

