% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{backtrace}
\alias{backtrace}
\title{back-tracing : collect clustered segments from the scoring function matrix}
\usage{
backtrace(S, K, multib, nextmax = FALSE, verb = TRUE)
}
\arguments{
\item{S}{matrix S, containing the local scores}

\item{K}{matrix K, containing the position k used for score maximization}

\item{multib}{if multiple k produce the maximal score, take either the shortest k ("max") or the longest k ("min"); TODO: what's the effect of "skip"?}

\item{nextmax}{go backwards while score is increasing before openening a new segment}

\item{verb}{print messages}
}
\description{
back-tracing : collect clustered segments from the scoring function matrix
}

