% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{calculateScoringMatrix}
\alias{calculateScoringMatrix}
\title{generate scoring function matrices for each cluster}
\usage{
calculateScoringMatrix(seq, C, score = "ccor", M, Mn, csim, ncpu = 1,
  preschedule = FALSE)
}
\arguments{
\item{seq}{a sequence of cluster assignments}

\item{C}{optional cluster sorting}

\item{score}{the scoring function to be used: ccor, icor or cls}

\item{M}{minimal sequence length; Note, that this is not a strict
cut-off but defined as a penalty that must be "overcome" by good score.}

\item{Mn}{minimal sequence length for nuissance cluster, Mn<M will allow
shorter distances between segments; only used in scoring functions
"ccor" and "icor"}

\item{csim}{cluster-cluster or position-cluster similarity
matrix, for scoring functions ccor and icor, respectively}

\item{ncpu}{number of available cores (CPUs), passed to
\code{parallel::mclapply}}

\item{preschedule}{\code{parallel::mclapply} option that currently
requires to be set to FALSE to avoid an error in data collection from
parallel processes}
}
\value{
Returns the scoring function matrices \code{SM} for all clusters
in the sequence \code{seq}.
}
\description{
generate scoring function matrices for each cluster
}

