% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateTotalScore}
\alias{calculateTotalScore}
\title{dynamic programming routine}
\usage{
calculateTotalScore(seq, C, SM, multi = "max", verb = 1L)
}
\arguments{
\item{seq}{the cluster sequence (where clusters at positions k:i are
considered).}

\item{C}{the list of clusters}

\item{SM}{list of scoring function matrices}

\item{multi}{if multiple \code{k} are found which return the same maximal
score, should the "max" (shortest distance) or "min" \code{k} be used?
This has little effect on real-life large data sets, since the situation
will rarely occur. Default is "max".}

\item{verb}{level of verbosity, currently not used (TODO: rm?)}
}
\value{
Returns the total score matrix \code{S(i,c)} and the matrix 
\code{K(i,c)} which stores the position \code{k} which delivered
the maximal score at position \code{i}. This is used in the back-tracing
phase.
}
\description{
dynamic programming routine
}
\details{
: This is \code{\link{segmenTier}}'s core dynamic programing
routine. It takes the scoring function matrices for all clusters,
 and dynamically constructs the total score matrix S(i,c).
}

