% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccSMccor}
\alias{ccSMccor}
\title{Scoring Function Matrix "ccor"}
\usage{
ccSMccor(seq, c, M, Mn, csim)
}
\arguments{
\item{seq}{the cluster sequence (where clusters at positions k:i are
considered). Note that \code{seq} is defined differently here than
the wrapper interfaces and MUST be a sequence of positive integers >0.}

\item{c}{the cluster to which similarities are to be calculated}

\item{M}{minimal sequence length; Note, that this is not a strict
cut-off but defined as a penalty that must be "overcome" by good score.}

\item{Mn}{minimal sequence length for nuissance cluster, Mn<M will allow
shorter distances between segments; only used in scoring functions
"ccor" and "icor"}

\item{csim}{cluster-cluster similarity matrix}
}
\value{
Returns the scoring matrix \code{SM(n,n)} for the cluster sequence
\code{seq} and cluster \code{c} for scoring function "ccor".
}
\description{
Scoring Function Matrix "ccor"
}
\details{
Scoring function "ccor" calculates the sum of similarities
between the clusters at positions k:i to cluster c over all k and i.
Note the difference to "icor" where real data from positions are
compared to cluster centers, while here two cluster centers are compared.
}

