% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccSMcls}
\alias{ccSMcls}
\title{Scoring Function Matrix "cls"}
\usage{
ccSMcls(seq, c, M, Mn, csim)
}
\arguments{
\item{seq}{the cluster sequence (where clusters at positions k:i are
considered). Note that \code{seq} is defined differently here than
in the wrapper interfaces and MUST be a sequence of positive integers >0.}

\item{c}{the cluster to which similarities are to be calculated}

\item{M}{minimal sequence length; Note, that this is not a strict
cut-off but defined as a penalty that must be "overcome" by good score.}

\item{Mn}{minimal sequence length for nuissance cluster, Mn<M will allow
shorter distances between segments; only used in scoring functions
"ccor" and "icor"}

\item{csim}{integer, the penalty \code{a} for non-matching clusters}
}
\value{
Returns the scoring matrix \code{SM(n,n)} for the cluster sequence
\code{seq} and cluster \code{c} for simplest scoring function "cls".
}
\description{
Scoring Function Matrix "cls"
}
\details{
Scoring function "cls" merely counts the number of
of clusters in sequence k:i, over all k and i, that are identical
to the tested cluster \code{c}, and sub-tracts a minimal size penality
and a penalty the for the count of non-identical clusters.
Note: this function used in the scoring unction scorecls_c for individual
calculations.
}

