% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccSMncor}
\alias{ccSMncor}
\title{like ccSMccor, but also calculates scores for the nuissance cluster}
\usage{
ccSMncor(seq, c, M, csim)
}
\arguments{
\item{seq}{the cluster sequence (where clusters at positions k:i are
considered)}

\item{c}{the cluster to which similarities are to be calculated}

\item{M}{initial penalty parameter, used to set minimal segment sizes}

\item{csim}{cluster-cluster similarity matrix}
}
\description{
like ccSMccor, but also calculates scores for the nuissance cluster
}

