% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clusterCor_c}
\alias{clusterCor_c}
\title{Calculates data-cluster similarity.}
\usage{
clusterCor_c(data, clusters)
}
\arguments{
\item{data}{original data matrix}

\item{clusters}{cluster centers}
}
\value{
Returns a data-cluster correlation matrix as used in
scoring function "icor".
}
\description{
Calculates data-cluster similarity.
}
\details{
calculates Pearson's product-moment correlation coefficients
of original data to cluster centers as used in the scoring function "icor".
}

