% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clusterMaxCor_c}
\alias{clusterMaxCor_c}
\title{Reports the maximally correlating cluster for each data point.}
\usage{
clusterMaxCor_c(data, clusters, mincor = 0, warn = 0L)
}
\arguments{
\item{data}{original data matrix}

\item{clusters}{cluster centers}

\item{mincor}{minimal correlation to be considered}

\item{warn}{if set to 1/TRUE then the occurence of multiple maxima
is tested and a warning issued}
}
\description{
Reports the maximally correlating cluster for each data point.
}
\details{
Calculates Pearson's product-moment correlation of each data
to clusters and report the (first!) cluster which had max correlation.
Note that this can lead to amibiguities since several clusters may
reach the same maximum.
}
