% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{coor2index}
\alias{coor2index}
\title{convert chromosome coordinates to continuous index}
\usage{
coor2index(features, chrS, cols = c("start", "end", "coor"), chrCol = "chr",
  strandCol = "strand", reverse = c("-", -1))
}
\arguments{
\item{features}{a table of chromosome features that must contain
the chromosome number (option \code{chrCol}), one or more chromosome
positions (option \code{cols}) and strand information (column
 \code{strandCol}).}

\item{chrS}{the chromosome index, indicating the start position
of each chromosome in the continuous index, derived from chromosome length
information}

\item{cols}{name of the columns giving coordinates that will be mapped
to continuous index}

\item{chrCol}{name of the column that gives the chromosome number}

\item{strandCol}{name of the column that gives forward/reverse strand
information}

\item{reverse}{a vector of possible reverse strand indicators}
}
\description{
convert chromosome coordinates to continuous index
}

