% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{fillScoringMatrix}
\alias{fillScoringMatrix}
\title{converts the vector form of the triangular scoring matrix
into a matrix}
\usage{
fillScoringMatrix(SV)
}
\arguments{
\item{SV}{vector form of the triangular scoring matrix as returned
by \code{\link{calculateScoringMatrix}}}
}
\description{
converts the vector form of the triangular scoring matrix
into a matrix
}

