% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{flowclusterTimeseries}
\alias{flowclusterTimeseries}
\title{Cluster a processed time-series with \pkg{flowClust} & \pkg{flowMerge}.}
\usage{
flowclusterTimeseries(tset, ncpu = 1, K = 10, merge = FALSE, B = 500,
  tol = 1e-05, lambda = 1, nu = 4, nu.est = 0, trans = 1, ...)
}
\arguments{
\item{tset}{processed time-series as provided by
\code{\link{processTimeseries}}}

\item{ncpu}{number of cores available for parallel mode of
\pkg{flowClust}. NOTE: parallel mode of
\code{\link[flowClust:flowClust]{flowClust}} is often non-functional.
For some installations the option \code{mc.cores=ncpu} can
be passed on to \code{\link[flowClust:flowClust]{flowClust}} to run
the algorithm in parallel mode.}

\item{K}{the requested cluster numbers (vector of integers)}

\item{merge}{logical indicating whether cluster merging with
\pkg{flowMerge} should be attempted}

\item{B}{max. num. of EM iterations}

\item{tol}{tolerance for EM convergence}

\item{lambda}{intial Box-Cox trafo}

\item{nu}{initial Box-Cox trafo, Inf for pure Gaussian}

\item{nu.est}{0: no, 1: non-specific, 2: cluster-specific estimation of nu}

\item{trans}{0: no, 1: non-specific, 2: cluster-specific estim. of lambda}

\item{...}{further parameter to \code{flowClust}}
}
\description{
A wrapper for \pkg{flowClust}, clustering
a time-series object \code{tset} provided by \code{\link{processTimeseries}},
where specifically the DFT of a time-series and requested data
transformation were calculated. The clustering is performed
on the matrix \code{tset$dat}. This is currently only used for
clustering of final segment time-series; it could in principle also
be used for segmentation, but that has not been tested.
Please see option \code{ncpu} on how to use parallel mode, which
does not work on some installations.
}
