% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_snr}
\alias{get_snr}
\title{Scale a DFT to a signal-to-noise ratio, where each frequency is divided
by the mean of all other frequencies.}
\usage{
get_snr(x)
}
\arguments{
\item{x}{a matrix of discrete fourier transformed (DFT) data, as provided
by \code{\link{get.fft}} or \code{\link[stats:fft]{mvfft}}.}
}
\description{
Scale a DFT to a signal-to-noise ratio, where each frequency is divided
by the mean of all other frequencies.
}
