% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{idx2chr}
\alias{idx2chr}
\title{get the chromosome from continuous index}
\usage{
idx2chr(idx, chrS)
}
\arguments{
\item{idx}{index position for which chromosome information is reported}

\item{chrS}{the chromosome index, indicating the start position
of each chromosome in the continuous index, derived from chromosome length
information}
}
\value{
returns the chromosome number
}
\description{
get the chromosome from continuous index
}

