% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{idx2coor}
\alias{idx2coor}
\title{Simple version of \code{\link{index2coor}} for single values}
\usage{
idx2coor(pos, chrS)
}
\arguments{
\item{pos}{the continuous index position that will be mapped to
chromosome coordinates}

\item{chrS}{the chromosome index, indicating the start position
of each chromosome in the continuous index, derived from chromosome length
information}
}
\description{
Simple version of \code{\link{index2coor}} for single values
}

