% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{idx2str}
\alias{idx2str}
\title{get the strand from continuous index}
\usage{
idx2str(idx, chrS)
}
\arguments{
\item{idx}{index position for which strand information is reported}

\item{chrS}{the chromosome index, indicating the start position
of each chromosome in the continuous index, derived from chromosome length
information}
}
\value{
returns the strand
}
\description{
get the strand from continuous index
}

