% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{index2coor}
\alias{index2coor}
\title{convert continuous index to chromosome coordinates (reverse of
\code{\link{coor2index}})}
\usage{
index2coor(features, chrS, cols = c("start", "end", "coor"), chrCol = "chr",
  strandCol = "strand", reverse = c("-", -1))
}
\arguments{
\item{features}{a table of chromosome features that must contain
the chromosome number (option \code{chrCol}), one or more chromosome
positions (option \code{coorCols}) and strand information (column
 \code{strandCol}).}

\item{chrS}{the chromosome index, indicating the start position
of each chromosome in the continuous index, derived from chromosome length
information}

\item{cols}{names of the columns giving coordinates that will be mapped
to continuous index}

\item{chrCol}{name of the column that gives the chromosome number}

\item{strandCol}{name of the column that gives forward/reverse strand
information}

\item{reverse}{a vector of possible reverse strand indicators}
}
\description{
convert continuous index to chromosome coordinates (reverse of
\code{\link{coor2index}})
}

