% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{log_1}
\alias{log_1}
\title{log transformation handling zeros by adding 1}
\usage{
log_1(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
A conventional approach to handle 0 in log transformation is to simply
add 1 to all data, \code{log_1(x) = log(x+1)}. Also see \code{\link{ash}}.
}
