% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.clustering}
\alias{plot.clustering}
\title{plot the clustering object returned by \code{\link{clusterTimeseries}}}
\usage{
\method{plot}{clustering}(x, k, xaxis, ...)
}
\arguments{
\item{x}{a clusterings object as returned by
\code{\link{clusterTimeseries}}}

\item{k}{a numeric or string vector indicating the clusterings to be plotted;
specifically the column numbers or names in the matrix of clusterings
in \code{cset$clusters}; if missing all columns will be plotted
and the calling code must take care of properly assigning \code{par(mfcol)}
or \code{layout} for the plot}

\item{xaxis}{optinally x-values to use as x-axis (e.g. to reflect absolute
chromosomal coordinates)}

\item{...}{currently unused additional arguments to plot}
}
\description{
plot the clustering object returned by \code{\link{clusterTimeseries}}
}

