% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{plot all objects from the segmentation pipeline, i.e. the processed
time-series, the clustering, the internal scoring matrices and
the final segments}
\usage{
plotSegmentation(tset, cset, sset)
}
\arguments{
\item{tset}{the time-series object returned by
\code{\link{processTimeseries}}}

\item{cset}{a set of clusterings as returned by
\code{\link{clusterTimeseries}}}

\item{sset}{a set of segmentations as returned by
\code{\link{segmentClusters}} and \code{\link{segmentCluster.batch}}}
}
\description{
plot all objects from the segmentation pipeline, i.e. the processed
time-series, the clustering, the internal scoring matrices and
the final segments
}

