% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{plotSegments}
\alias{plotSegments}
\title{plot segmentation data based on data returned by
high-level wrappers, incl. S(c,i), K(c,i) and segments}
\usage{
plotSegments(scrR, seq, ts, tot, out, use.log = FALSE, add.plots = 0,
  verb = 2)
}
\description{
plot segmentation data based on data returned by
high-level wrappers, incl. S(c,i), K(c,i) and segments
}

