% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{print.segments}
\alias{print.segments}
\title{Print method for segmentation result from \code{\link{segmentClusters}}.}
\usage{
\method{print}{segments}(x, ...)
}
\arguments{
\item{x}{result object returned by function \code{\link{segmentClusters}}}

\item{...}{further argument to \code{print.data.frame}}
}
\description{
Print method for segmentation result from \code{\link{segmentClusters}}.
}
