% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scorecls_c}
\alias{scorecls_c}
\title{Scoring Function "cls"}
\usage{
scorecls_c(k, i, c, seq, M, a)
}
\arguments{
\item{k}{start position for score calculation}

\item{i}{end position for score calculation}

\item{c}{the cluster to which similarities are to be calculated}

\item{seq}{the cluster sequence (where clusters at positions k:i are
considered)}

\item{M}{initial penalty parameter, used to set minimal segment sizes}

\item{a}{the penalty for non-matching clusters}
}
\description{
Scoring Function "cls"
}
\details{
Scoring function "cls" merely counts the number of
of clusters in sequence k:i that are identical to the tested
cluster \code{c}, and sub-tracts a minimal size penality
and a penalty the for the count of non-identical clusters.
NOTE: This function is used in the matrix function ccSMcls.
}

