% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scorecor_c}
\alias{scorecor_c}
\title{Scoring Function "cor" - Test}
\usage{
scorecor_c(k, i, c, seq, M, csim)
}
\arguments{
\item{k}{start position for score calculation}

\item{i}{end position for score calculation}

\item{c}{the cluster to which similarities are to be calculated}

\item{seq}{the cluster sequence (where clusters at positions k:i are
considered)}

\item{M}{initial penalty parameter, used to set minimal segment sizes}

\item{csim}{cluster-cluster similarity matrix}
}
\description{
Scoring Function "cor" - Test
}
\details{
Scoring function "cor" calculates the sum of similarities
between the clusters at positions k:i to cluster c. Note the difference
to "icor" where real data from positions are comapred to clusters, while
here two clusters are compared.
NOTE: This function is for testing only, the dynamic programming algorithm 
uses function ccSMcor.
}

