% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{segment.plotFeatures}
\alias{segment.plotFeatures}
\title{plot features as arrows}
\usage{
segment.plotFeatures(data, coors, types, strand, typord = FALSE,
  cuttypes = FALSE, names = FALSE, legend = FALSE, axis1 = FALSE,
  ylab = NA, args, line = 1, ycx = 1, tcx = 1, tpos = NULL,
  columns = c(name = "name", chr = "chr", strand = "strand", start = "start",
  end = "end", type = "type", color = "color"))
}
\description{
plot features as arrows
}

