% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{segment.plotHeat}
\alias{segment.plotHeat}
\title{plot 3D genomeData as a heatmap}
\usage{
segment.plotHeat(data, coors, orig.idx, breaks, colors, orig.col, ylab = "",
  axes = FALSE, axis2 = FALSE, colnorm = FALSE, chrS, coor, args)
}
\description{
plot 3D genomeData as a heatmap
}

