% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{segmentCluster.batch}
\alias{segmentCluster.batch}
\title{A high-level wrapper for multiple runs of segmentation by
\code{\link{segmentClusters}} for multiple clusterings and
multiple segmentation parameters. It additionally allows to
tag adjacent segments to be potentially fused due to similarity
of their clusters.}
\usage{
segmentCluster.batch(cset, fuse.threshold = 0.2, short.name = TRUE, id,
  ncpu = 1, verb = 1, save.matrix = FALSE, varySettings = list(E = 1, S
  = "ccor", M = c(100), Mn = c(20, 100), a = -2, nui = 1, nextmax = TRUE, multi
  = "max", multib = "max"))
}
\arguments{
\item{cset}{a clustering set as returned by \code{\link{clusterTimeseries}}}

\item{fuse.threshold}{if adjacent segments are associated with clusters
the centers of which have a Pearson correlation \code{>fuse.threshold}
the field "fuse" will be set to 1 for the second segments (top-to-bottom
as reported)}

\item{short.name}{if TRUE (default) parameters that are not varied
will not be part of the segment type and ID}

\item{id}{if set, the default segment IDs, constructed from numbered
segment types, are replaced by this}

\item{ncpu}{number of available cores (CPUs), passed to
\code{\link[parallel:mclapply]{parallel::mclapply}} by
\code{\link{calculateScoringMatrix}}}

\item{verb}{level of verbosity, 0: no output, 1: progress messages}

\item{save.matrix}{store the total score matrix \code{S(i,c)} and the
backtracing matrix \code{K(i,c)}; useful in testing stage or for
debugging or illustration of the algorithm; see \code{\link{plotScoring}}
TODO: save.matrix is currently not implemented, since batch function
returns a matrix only}

\item{varySettings}{list of settings where each entry can be a vector;
the function will construct a matrix of all possible combinations of
parameter values in this list, call \code{\link{segmentClusters}} for
each, and report a matrix of segments where the segment `type' is
constructed from the varied parameters; see option \code{short.name}}
}
\description{
A high-level wrapper for multiple runs of segmentation by
\code{\link{segmentClusters}} for multiple clusterings and
multiple segmentation parameters. It additionally allows to
tag adjacent segments to be potentially fused due to similarity
of their clusters.
}
\details{
This is a high-level wrapper for \code{\link{segmentClusters}}
which allows segmentation over multiple clusterings as provided by the
function \code{\link{clusterTimeseries}} and over multiple segmentation
paramers. Specifially, parameters \code{csim.scale}, \code{score},
\code{M} and \code{Mn} can all be vectors.
}

