% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segmentClusters}
\alias{segmentClusters}
\title{segmenTier - main wrapper interface, to be used in commandline script}
\usage{
segmentClusters(seq, csim, score = "ccor", M = 175, Mn = 20, a = 2,
  nui = 1, SM, multi = "min", multib = "min", nextmax = TRUE,
  ncpu = 1, verb = 1, save.mat = "")
}
\arguments{
\item{seq}{a clustering sequence}

\item{csim}{cluster-cluster or position-cluster similarity
matrix, for scoring functions ccor and icor, respectively}

\item{score}{the scoring function to be used: "ccor", "icor" or "cls"}

\item{M}{penalty for short sequences}

\item{Mn}{penalty for nuissance cluster, Mn<M will allow shorter distances
between segments}

\item{multi}{handling of multiple k with max. score in forward phase, see
Details. Either "min" (default) or "max"}

\item{multib}{handling of multiple k with max. score in back-trace phase, see
Details. Either "min" (default), "max" or "skip"}

\item{nextmax}{go backwards while score is increasing before openening a
new segment, default is TRUE}

\item{save.mat}{store matrices S(c,i) and K(c,i) for plotting and inspection,
mostly useful for debugging or illustration of the algorithm}
}
\description{
segmenTier - main wrapper interface, to be used in commandline script
}
\details{
This is the main R wrapper function for the segmentation algorithm.
It takes a sequence of clusterings and returns segments of consistent
clusters. Some more details of the algorithm can be tuned, but these usually
have no effect on real-life data sets. If multiple k produce the maximal
score, take either the shortest k ("max") or the longest k ("min");
TODO: what's the effect of "skip"?
}

