% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\docType{package}
\name{segmenTier}
\alias{segmenTier-package}
\alias{segmenTier}
\title{segmenTier : cluster-based segmentation
from a sequential clustering}
\description{
A dynamic programming solution to segmentation based on maximization of arbitrary similarity measures within segments. The general idea, theory and this implementation are described in Machne, Murray & Stadler (2017) \doi{10.1038/s41598-017-12401-8}. In addition to the core algorithm, the package provides time-series processing and clustering functions as described in the publication. These are generally applicable where a `k-means` clustering yields meaningful results, and have been specifically developed for clustering of the Discrete Fourier Transform of periodic gene expression data (`circadian' or `yeast metabolic oscillations'). This clustering approach is outlined in the supplemental material of Machne & Murray (2012) \doi{10.1371/journal.pone.0037906}), and here is used as a basis of segment similarity measures. Notably, the time-series processing and clustering functions can also be used as stand-alone tools, independent of segmentation, e.g., for transcriptome data already mapped to genes.
}
\section{Dependencies}{
 The package strictly depends only on
\code{Rcpp}.
All other dependencies are usually present in a
basic installation (\code{stats}, \code{graphics}, \code{grDevices})).
}

\references{
Machne, Murray & Stadler (2017) <doi:10.1038/s41598-017-12401-8>,
Machne & Murray (2012) <doi:10.1371/journal.pone.0037906>, and
Lehmann et al. (2013) <doi:10.1186/1471-2105-14-133>
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/raim/segmenTier}
  \item Report bugs at \url{https://github.com/raim/segmenTier/issues}
}

}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}, Douglas B. Murray,
Peter F. Stadler \email{studla@bioinf.uni-leipzig.de}
}
