% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{setVarySettings}
\alias{setVarySettings}
\title{generate the parameter list (\code{varySettings}) for
\code{\link{segmentCluster.batch}}, using defaults
for all parameters not passed.}
\usage{
setVarySettings(E = 1, S = "ccor", M = c(100), Mn = c(20, 100),
  a = -2, nui = 1, nextmax = TRUE, multi = "max", multib = "max")
}
\description{
generate the parameter list (\code{varySettings}) for
\code{\link{segmentCluster.batch}}, using defaults
for all parameters not passed.
}

