% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{sortClusters}
\alias{sortClusters}
\title{takes a clustering set as returned by \code{\link{clusterTimeseries}} and
uses the cluster-cluster similarity matrix \code{Ccc} to re-sort
clusters by their similarity, starting with the first cluster; the
sorting is added as \code{sorting} to \code{cset} and the annotated
\code{cset} is returned.}
\usage{
sortClusters(cset, verb = 0)
}
\arguments{
\item{cset}{a clustering set as returned by \code{\link{clusterTimeseries}}}

\item{verb}{level of verbosity, 0: no output, 1: progress messages}
}
\description{
takes a clustering set as returned by \code{\link{clusterTimeseries}} and
uses the cluster-cluster similarity matrix \code{Ccc} to re-sort
clusters by their similarity, starting with the first cluster; the
sorting is added as \code{sorting} to \code{cset} and the annotated
\code{cset} is returned.
}

