% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{testPhase}
\alias{testPhase}
\title{tests phase recovery by \code{\link{phase}}}
\usage{
testPhase(n = 5, cyc = 4, T = 24, res = 6, xlim = c(-T/res, T +
  T/res))
}
\arguments{
\item{n}{number of cosines to generate}

\item{cyc}{numbers of cycles to generate}

\item{T}{period}

\item{res}{resolution: samples per period}

\item{xlim}{x-axis range to show, defaults to one period}
}
\description{
tests phase recovery by \code{\link{phase}}
}
