% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{writeSegments}
\alias{writeSegments}
\title{writing out data for segments to files, used for writing the primary
segments by \code{\link{presegment}} to single files that are
then further processed by segmenTier.}
\usage{
writeSegments(data, segments, name, path)
}
\arguments{
\item{data}{a data matrix to which coordinates in \code{segments}
refer to}

\item{segments}{a matrix that must contain "start" and "end" (columns)
of segments; these coordinates will be extracted from \code{data}
and written to individual files, numbered by the row number in segments.}

\item{path}{optional output path where files will be written, if not supplied
files will end up in the current working directory (`getwd`)}
}
\description{
writing out data for segments to files, used for writing the primary
segments by \code{\link{presegment}} to single files that are
then further processed by segmenTier.
}

