% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{ash}
\alias{ash}
\title{\code{asinh} data transformation}
\usage{
ash(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
The asinh transformation, (\code{ash(x) = log(x + sqrt(x^2+1))}), is
an alternative to log transformation that has less (compressing) effects
on the extreme values (low and high values), and naturally handles
negative numbers and 0. Also see \code{\link{log_1}}.
}
