% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{calculateScoringMatrix}
\alias{calculateScoringMatrix}
\title{generate scoring function matrices for each cluster}
\usage{
calculateScoringMatrix(seq, C, score = "ccor", M, Mn, csim, ncpu = 1,
  preschedule = FALSE)
}
\arguments{
\item{seq}{a sequence of cluster assignments}

\item{C}{optional cluster sorting}

\item{score}{the scoring function to be used: ccor, icor or cls}

\item{M}{penalty for short sequences}

\item{Mn}{penalty for nuissance cluster, Mn<M will allow shorter distances
between segments}

\item{csim}{cluster-cluster or position-cluster similarity
matrix, for scoring functions ccor and icor, respectively}

\item{ncpu}{number of available cores (CPUs), passed to
\code{parallel::mclapply}}

\item{preschedule}{\code{parallel::mclapply} option that currently
requires to be set to FALSE to avoid an error in data collection from
parallel processes}
}
\description{
generate scoring function matrices for each cluster
}

