% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccSMccor}
\alias{ccSMccor}
\title{Scoring Function Matrix "ccor"}
\usage{
ccSMccor(seq, c, M, Mn, csim)
}
\arguments{
\item{seq}{the cluster sequence (where clusters at positions k:i are
considered)}

\item{c}{the cluster to which similarities are to be calculated}

\item{M}{initial penalty parameter, used to set minimal segment sizes}

\item{Mn}{penalty parameter for nuissance clusters, should be Mn<M}

\item{csim}{cluster-cluster similarity matrix}
}
\description{
Scoring Function Matrix "ccor"
}
\details{
Scoring function "ccor" calculates the sum of similarities
between the clusters at positions k:i to cluster c over all k and i.
Note the difference to "icor" where real data from positions are
compared to clusters, while here two clusters are compared.
}

