% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccSMcls}
\alias{ccSMcls}
\title{Scoring Function Matrix "cls"}
\usage{
ccSMcls(seq, c, M, Mn, csim)
}
\arguments{
\item{seq}{the cluster sequence (where clusters at positions k:i are
considered)}

\item{c}{the cluster to which similarities are to be calculated}

\item{M}{initial penalty parameter, used to set minimal segment sizes}

\item{Mn}{not use here, present just for consistency between
scoring matrix function calls}

\item{csim}{integer, the penalty for non-matching clusters}
}
\description{
Scoring Function Matrix "cls"
}
\details{
Scoring function "cls" merely counts the number of
of clusters in sequence k:i, over all k and i, that are identical
to the tested cluster \code{c}, and sub-tracts a minimal size penality
and a penalty the for the count of non-identical clusters.
Note: this function used in the scoring unction scorecls_c for individual
calculations.
}

