% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccSMicor}
\alias{ccSMicor}
\title{Scoring Function Matrix "icor"}
\usage{
ccSMicor(seq, c, M, Mn, csim)
}
\arguments{
\item{seq}{the cluster sequence (where positions k:i are considered);
notably this argument not required here, but only used for
consistency with other scoring functions}

\item{c}{the cluster to which similarities are to be calculated; note, 
that c=1 is the nuissance cluster}

\item{M}{initial penalty parameter, used to set minimal segment sizes}

\item{Mn}{penalty parameter for nuissance clusters, should be Mn<M}

\item{csim}{position-cluster similarity matrix, where the rows
are the positions in the sequence \code{seq} and columns are the
the clusters}
}
\description{
Scoring Function Matrix "icor"
}
\details{
Scoring function "icor" calculates the sum of similarities of
positions k:i to cluster c over all k and i.
The similarities are calculated e.g., as a (Pearson) correlation between
the individual positions and the tested cluster c center.
Note the difference to "ccor" where the cluster centers are compared
instead of original data at positions k and i with a cluster.
}

