% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccSMncor}
\alias{ccSMncor}
\title{like ccSMccor, but also calculates scores for the nuissance cluster}
\usage{
ccSMncor(seq, c, M, Mn, csim)
}
\arguments{
\item{seq}{the cluster sequence (where clusters at positions k:i are
considered). Note that \code{seq} is defined differently here than
the wrapper interfaces and MUST be a sequence of positive integers >0.}

\item{c}{the cluster to which similarities are to be calculated}

\item{M}{minimal sequence length; Note, that this is not a strict
cut-off but defined as a penalty that must be "overcome" by good score.}

\item{Mn}{minimal sequence length for nuissance cluster, Mn<M will allow
shorter distances between segments; only used in scoring functions
"ccor" and "icor"}

\item{csim}{cluster-cluster similarity matrix}
}
\description{
like ccSMccor, but also calculates scores for the nuissance cluster
}

