% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clusterCor_c}
\alias{clusterCor_c}
\title{Calculates position-cluster correlations for scoring function "icor".}
\usage{
clusterCor_c(data, clusters)
}
\arguments{
\item{data}{original data matrix}

\item{clusters}{cluster centers}
}
\value{
Returns a position-cluster correlation matrix as used in
scoring function "icor".
}
\description{
Calculates Pearson's product-moment correlation coefficients
between rows in \code{data}  and \code{cluster}, and is used to calculate
the position-cluster similarity matrix for the scoring function "icor".
This is implemented in Rcpp for calculation speed, using 
\code{\link{myPearson}} to calculate correlations.
}
