% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterTimeseries}
\alias{clusterTimeseries}
\title{wrapper for \code{\link[stats:kmeans]{kmeans}} clustering
of a time-series preprocessed by \code{\link{processTimeseries}}.}
\usage{
clusterTimeseries(tset, K = 16, iter.max = 1e+05, nstart = 100,
  nui.thresh = -Inf, verb = 1)
}
\arguments{
\item{tset}{a timeseries processed by \code{\link{processTimeseries}}}

\item{K}{selected cluster numbers, the argument \code{centers}
of \code{\link[stats:kmeans]{kmeans}}}

\item{iter.max}{the maximum number of iterations allowed in
\code{\link[stats:kmeans]{kmeans}}, see there}

\item{nstart}{initialization \code{\link[stats:kmeans]{kmeans}}:
"how many random sets should be chosen?", see there}

\item{nui.thresh}{threshold correlation of a data point to a cluster
center; if below the data point will be added to nuissance cluster 0}

\item{verb}{level of verbosity, 0: no output, 1: progress messages}
}
\description{
wrapper for \code{\link[stats:kmeans]{kmeans}} clustering
of a time-series preprocessed by \code{\link{processTimeseries}}.
}

