% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{colorClusters}
\alias{colorClusters}
\title{Assign colors to clusters.}
\usage{
colorClusters(cset)
}
\arguments{
\item{cset}{a clustering set as returned by \code{\link{clusterTimeseries}}}
}
\description{
Takes a clustering set as returned by \code{\link{clusterTimeseries}} and
assigns colors to each cluster in each clustering along
the "hue" color wheel, as in \code{scale_colour_hue} in \code{ggplot2};
if \code{cset} contains a sorting (see
\code{clusterSort}), this sorting will be used to assing colors along
the color wheel, otherwise a sorting will be calculated first;
}
