% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{logLik.kmeans}
\alias{logLik.kmeans}
\title{Experimental: AIC/BIC for kmeans}
\usage{
\method{logLik}{kmeans}(object, ...)
}
\arguments{
\item{object}{a \code{\link[stats:kmeans]{kmeans}} result object}

\item{...}{unused}
}
\description{
This function is supposed to provide a log-likelihood method for
\code{\link[stats:kmeans]{kmeans}} results, after Neal Fultz at
\url{https://stackoverflow.com/a/33202188} and also featured in the
\href{https://rdrr.io/github/nfultz/stackoverflow/src/R/logLik_kmeans.R}{stackoverflow package}. Note, that the blogged version on Jan 30, 2019 adds
a minus and a division by 2 compared to a linked git version.
This idea has not been reviewed, and this function has not been
tested extensively; feel free to do so and contribute your results.
}
\details{
This is an attempt to reproduce the \code{BIC} measure
in model-based clustering to decide on an optimal number of clusters.
This function will be used for \code{\link[stats:kmeans]{kmeans}}
results objects when passed to \code{\link[stats:BIC]{BIC}} and
\code{\link[stats:AIC]{AIC}} functions from the \pkg{stats} package in
base R, and BIC and AIC are calculated this way in
\code{\link{segmentClusters}}. It is however not used anywhere at the
moment.
}
