% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{myPearson}
\alias{myPearson}
\title{Pearson product-moment correlation coefficient}
\usage{
myPearson(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
Incremental calculation of the Pearson correlation coefficient between
two vectors for calculation within Rcpp functions
\code{\link{clusterCor_c}}.
}
\details{
Simply calculates Pearson's product-moment correlation
between vectors \code{x} and \code{y}.
}
