% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{phase}
\alias{phase}
\title{calculate phase
calculates the phases (in degrees: 0-360) via
a Discrete Fourier Transformation}
\usage{
phase(x, cycles, degrees = TRUE)
}
\arguments{
\item{x}{a time-series matrix with columns as time points,
or a \code{timeseries} object as returned by \code{\link{processTimeseries}}}

\item{cycles}{optional integer vector for the DFT components
(number of cycles in the data) for which phases are to be calculated;
all are returned if \code{cycle} is not specified}

\item{degrees}{logical to indicate whether phases should be reported
in degrees (0-360) or radians (-pi - +pi)}
}
\description{
calculate phase
calculates the phases (in degrees: 0-360) via
a Discrete Fourier Transformation
}
