% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.segments}
\alias{plot.segments}
\title{plot the final segmentation objects returned by
\code{\link{segmentClusters}} and \code{\link{segmentCluster.batch}}}
\usage{
\method{plot}{segments}(x, types, xaxis, plot = c("segments", "S", "S1"), ...)
}
\arguments{
\item{x}{a set of segmentations as returned by
\code{\link{segmentClusters}} and \code{\link{segmentCluster.batch}}}

\item{types}{a string vector indicating segment types to plot (a subset of
\code{sset$ids}; defaults to all in \code{sset$ids})}

\item{xaxis}{optional x-values to use as x-axis (e.g. to reflect absolute
chromosomal coordinates)}

\item{plot}{string list indicating which data should be plotted;
`segments': plot segments as arrows; `S1' plot the scoring vectors
\code{s(i,j,c} for all \code{c}; `S' plot the derivative of
matrix \code{S(i,c)} for all \code{c}}

\item{...}{currently unused additional arguments to plot}
}
\description{
plot the final segmentation objects returned by
\code{\link{segmentClusters}} and \code{\link{segmentCluster.batch}}
}

