% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotScoring}
\alias{plotScoring}
\title{plot the scoring function matrices as a heatmap}
\usage{
plotScoring(SML, seq, score, out.file, verb = 2)
}
\arguments{
\item{SML}{a list with scoring function matrices for each cluster,
as provided by \code{\link{calculateScoringMatrix}}}

\item{seq}{the original cluster sequence, optional for axis labeling}

\item{score}{name of the used scoring function}

\item{out.file}{if supplied the scoring matrices will be plotted to
individual png files named <out.file>_<number>.png}

\item{verb}{level of verbosity; 0: no output, 1: progress messages}
}
\description{
plot the scoring function matrices as a heatmap
}

