% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Summary plot for the \code{segmenTier} pipeline.}
\usage{
plotSegmentation(tset, cset, sset, split = FALSE, plot.matrix = FALSE,
  mai = c(0.01, 2, 0.01, 0.01), ...)
}
\arguments{
\item{tset}{a time-series object as returned by
\code{\link{processTimeseries}}}

\item{cset}{a clusterings object as returned by
\code{\link{clusterTimeseries}}}

\item{sset}{a segmentation object as returned by
\code{\link{segmentClusters}} and \code{\link{segmentCluster.batch}}}

\item{split}{split segment plots by clustering plots}

\item{plot.matrix}{include the internal scoring matrices in the plot}

\item{mai}{margins of individual plots, see \code{par}}

\item{...}{further arguments to
\code{\link{plot.clustering}} (\code{cset}) and
\code{\link{plot.segments}} (\code{sset}). Note: these may
conflict and cause errors, but eg. a combination of \code{cex=0.5, lwd=3}
works, affecting cluster point size and segment line width, respectively.}
}
\description{
Plot all objects from the segmentation pipeline, i.e. the processed
time-series, the clustering, the internal scoring matrices and
the final segments.
}
