% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{plot all objects from the segmentation pipeline, i.e. the processed
time-series, the clustering, the internal scoring matrices and
the final segments}
\usage{
plotSegmentation(tset, cset, sset, plot.matrix = FALSE, mai = c(0.01, 1.5,
  0.01, 0.01), ...)
}
\arguments{
\item{tset}{a time-series object as returned by
\code{\link{processTimeseries}}}

\item{cset}{a clusterings object as returned by
\code{\link{clusterTimeseries}}}

\item{sset}{a segmentation object as returned by
\code{\link{segmentClusters}} and \code{\link{segmentCluster.batch}}}

\item{plot.matrix}{include the internal scoring matrices in the plot}

\item{mai}{margins of invidual plots, see \code{par}}
}
\description{
plot all objects from the segmentation pipeline, i.e. the processed
time-series, the clustering, the internal scoring matrices and
the final segments
}

