% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSegments}
\alias{plotSegments}
\title{plot segmentation data based on data returned by
high-level wrappers, incl. S(c,i), K(c,i) and segments}
\usage{
plotSegments(cset, ts, scrR, tot, out.file, use.log = FALSE, add.plots = 0,
  verb = 2)
}
\arguments{
\item{ts}{an optional time-series from for which the clustering was
calculated}

\item{scrR}{a list containing multiple segmentations for (1) different
scoring functions, (2) different parameters of the forward-step (dyn.prog.)
and (3) different parameters of the back-tracing step.}

\item{tot}{an optional total signal from the timeseries in \code{ts}}

\item{out.file}{optional out.file name (w/o file extension) to which a
png will be plotted}

\item{use.log}{plot the total data (\code{tot}) with logged y-axis}

\item{add.plots}{in the plot layout, leave these rows for additional
external plots}

\item{verb}{level of verbosity; 0: no output, 1: progress messages}

\item{seq}{the sequence of clusters to be segments}
}
\value{
Returns the file.name if out.file was specified.
}
\description{
plot segmentation data based on data returned by
high-level wrappers, incl. S(c,i), K(c,i) and segments
}
\details{
This is mostly used for testing, where a small data set
is segmented by a high-level wrapper for multiple scoring functions
and parameters.
}

