% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{processTimeseries}
\alias{processTimeseries}
\title{Process a time-series for \code{\link{segmenTier}}.}
\usage{
processTimeseries(ts, trafo = "raw", use.fft = TRUE, dc.trafo = "raw",
  dft.range, perm = 0, use.snr = TRUE, lambda = 1, low.thresh = -Inf,
  smooth.space = 1, smooth.time = 1, circular.time = FALSE, verb = 0)
}
\arguments{
\item{ts}{the timeseries as a matrix, where columns are the timepoints
and rows individual measurements (e.g., genomic positions for transcriptome
data)}

\item{trafo}{prior data transformation, pass any function name, e.g.,
"log", or the package functions "ash" (\code{asinh = ln(x + sqrt(x^2+1))})
or "log_1" for (\code{ln(ts+1)})}

\item{use.fft}{use the Discrete Fourier Transform of the data}

\item{dc.trafo}{data transformation for the first (DC) component of
the DFT, pass any function name, e.g., "log", or the package functions
"ash" (\code{asinh= ln(x + sqrt(x^2+1))}) or "log_1" for (\code{ln(ts+1)}).}

\item{dft.range}{a vector of integers, giving the components of the
Discrete Fourier Transform to be used where 1 is the first component (DC)
corresponding to the mean value, and 2:n are the higher components
correspondong to 2:n full cycles in the data}

\item{perm}{number of permutations of the data set, to obtain}

\item{use.snr}{use a scaled amplitude, where each component of the
Discrete Fourier Transform is divided by the mean of all other components,
which is similar to a signal-to-noise ratio (SNR)}

\item{lambda}{parameter lambda for Box-Cox transformation of
DFT amplitudes (experimental; not tested)}

\item{low.thresh}{use this threshold to cut-off data, which will be
added to the absent/nuissance cluster later}

\item{smooth.space}{integer, if set a moving average is calculated for
each time-point between adjacent data points using stats
package's \code{link[stats:smooth]{smooth}} with span \code{smooth.space}}

\item{smooth.time}{integer, if set the time-series will be smoothed
using stats package's \code{link[stats:filter]{filter}} to calculate a
moving average with span \code{smooth.time} and
\code{link[stats:smoothEnds]{smoothEnds}} to extrapolate smoothed first
and last time-points (again using span \code{smooth.time})}

\item{circular.time}{logical value indicating whether time can be treated
as circular in smoothing}

\item{verb}{level of verbosity, 0: no output, 1: progress messages}
}
\description{
Performs the requested data transformations, including a Discrete
Fourier Transform (DFT) of the time-series as direct input for 
the clustering wrapper \code{\link{clusterTimeseries}}.
}
\details{
This function exemplifies the processing of an oscillatory
transcriptome time-series data as used in the establishment of this
algorithm and the demo \code{segment_test}. As suggested by Machne & Murray
(PLoS ONE 2012) and Lehmann et al. (BMC Bioinformatics 2014) a Discrete
Fourier Transform of time-series data allows to cluster time-series by
their change pattern.
}
\references{
@cite Machne2012 Lehmann2013
}
