% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{processTimeseries}
\alias{processTimeseries}
\title{process a time-series apt for the \code{\link{segmenTier}}
clustering wrapper \code{\link{clusterTimeseries}}}
\usage{
processTimeseries(ts, smooth = FALSE, trafo = "identity",
  keep.zeros = FALSE, use.fft = TRUE, dc.trafo = "identity",
  dft.range = 2:7, use.snr = TRUE, low.thresh = 1)
}
\arguments{
\item{ts}{the timeseries as a matrix, where columns are the timepoints
and rows individual measurements (e.g., genomic positions for transcriptome
data)}

\item{smooth}{use stats' package \code{link[stats:smooth]{smooth}} to
smooth timeseries before processing}

\item{trafo}{prior data transformation, pass any function name, e.g.,
"log", or the package functions "asinh" (\code{ln(x + sqrt(x^2+1))}) or
"log_1" for (\code{ln(ts+1)})}

\item{use.fft}{use the Discrete Fourier Transform of the data}

\item{dc.trafo}{data transformation for the first (DC) component of
the DFT, pass any function name, e.g., "log", or the package functions
"asinh" (\code{ln(x + sqrt(x^2+1))}) or "log_1" for (\code{ln(ts+1)})}

\item{dft.range}{a vector of integers, giving the components of the
Discrete Fourier Transform to be used where 1 is the first component (DC)
corresponding to the mean value, and 2:n are the higher components
correspondong to 2:n full cycles in the data}

\item{use.snr}{use a scaled amplitude, where each component of the
Discrete Fourier Transform is divided by the mean of all other components,
which is similar to a signal-to-noise ratio (SNR)}

\item{low.thresh}{use this threshold to cut-off data, which will be
added to the absent/nuissance cluster later}
}
\description{
process a time-series apt for the \code{\link{segmenTier}}
clustering wrapper \code{\link{clusterTimeseries}}
}
\details{
This function exemplifies the processing of an oscillatory
transcriptome time-series data as used in the establishment of this
algorithm and the demo \code{segment_test}. As suggested by Machne & Murray
(PLoS ONE 2012) and Lehmann et al. (BMC Bioinformatics 2014) a Discrete
Fourier Transform of time-series data allows to cluster time-series by
their change pattern.
}
\references{
@cite Machne2012 Lehmann2013
}

