% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scoreicor_c}
\alias{scoreicor_c}
\title{Scoring Function "icor" - Test}
\usage{
scoreicor_c(k, i, c, seq, M, csim)
}
\arguments{
\item{k}{start position for score calculation}

\item{i}{end position for score calculation}

\item{c}{the cluster to which similarities are to be calculated}

\item{seq}{the cluster sequence (where positions k:i are considered);
notably this is not required here, but used as an argument for
consistency with other scoring functions.}

\item{M}{initial penalty parameter, used to set minimal segment sizes}

\item{csim}{position-cluster similarity matrix, where the rows
are the positions in the sequence \code{seq} and columns are the
the clusters}
}
\description{
Scoring Function "icor" - Test
}
\details{
Scoring function "icor" calculates the sum of similarities of
positions k:i to cluster c.
The similarities are calculated e.g., as a (Pearson) correlation between
the individual positions and the tested cluster c center. 
This function is for testing only, the dynamic programming algorithm 
uses function ccSMicor.
NOTE: individual scoring functions are only used for testing
the algorithm uses the matrix-filling functions below
}

