% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{segmentCluster.batch}
\alias{segmentCluster.batch}
\title{high-level wrapper for multiple runs of segmentation by
\code{\link{segmentClusters}} for multiple clusterings and
multiple segmentation parameters}
\usage{
segmentCluster.batch(cset, csim.scale = 1, score = "ccor", M = 175,
  Mn = 20, a = -2, nui = 1, fuse.threshold = 0.2, nextmax = TRUE,
  multi = "max", multib = "max", ncpu = 1, verb = 1,
  short.name = TRUE, save.mat = "")
}
\arguments{
\item{cset}{a clustering set as returned by \code{\link{clusterTimeseries}}}

\item{csim.scale}{exponent to scale similarity matrices, must be odd
to maintain negative correlations!}

\item{score}{the scoring function to be used: "ccor", "icor" or "cls"}

\item{M}{minimal sequence length; Note, that this is not a strict
cut-off but defined as a penalty that must be "overcome" by good score.}

\item{Mn}{minimal sequence length for nuissance cluster, Mn<M will allow
shorter distances between segments; only used in scoring functions
"ccor" and "icor"}

\item{a}{an additional penalty only used for pure cluster-based
scoring w/o cluster similarity measures in scoring function "cls"}

\item{nui}{the similarity score to be used for nuissance clusters in the
cluster similarity matrices}

\item{fuse.threshold}{if adjacent segments are associated with clusters
the centers of which have a Pearson correlation \code{>fuse.threshold}
the field "fuse" will be set to 1 for the second segments (top-to-bottom
as reported)}

\item{nextmax}{go backwards while score is increasing before openening a
new segment, default is TRUE}

\item{multi}{handling of multiple k with max. score in forward phase,
either "min" (default) or "max"}

\item{multib}{handling of multiple k with max. score in back-trace phase,
either "min" (default), "max" or "skip"}

\item{ncpu}{number of available cores (CPUs), passed to
\code{\link[parallel:mclapply]{parallel::mclapply}} by
\code{\link{calculateScoringMatrix}}}

\item{verb}{level of verbosity, 0: no output, 1: progress messages}

\item{short.name}{if TRUE (default) parameters that are not varied
will not be part of the segment type and ID}

\item{save.mat}{store the scoring function matrix SM or the back-tracing
matrix K by adding "SM" and "SK" to the string vector save.mat; useful
in testing stage or for debugging or illustration of the algorithm;
see \code{\link{plotScoring}}}
}
\description{
high-level wrapper for multiple runs of segmentation by
\code{\link{segmentClusters}} for multiple clusterings and
multiple segmentation parameters
}
\details{
This is a high-level wrapper for \code{\link{segmentClusters}}
which allows segmentation over multiple clusterings as provided by the
function \code{\link{clusterTimeseries}} and over multiple segmentation
paramers. Specifially, parameters \code{csim.scale}, \code{score},
\code{M} and \code{Mn} can all be vectors.
}

