% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{segmentCluster.batch}
\alias{segmentCluster.batch}
\title{Batch wrapper for \code{\link{segmentClusters}}.}
\usage{
segmentCluster.batch(cset, fuse.threshold = 0.2, short.name = TRUE, id,
  save.matrix = FALSE, verb = 1, varySettings = setVarySettings())
}
\arguments{
\item{cset}{a clustering set as returned by \code{\link{clusterTimeseries}}}

\item{fuse.threshold}{if adjacent segments are associated with clusters
the centers of which have a Pearson correlation \code{>fuse.threshold}
the field "fuse" will be set to 1 for the second segments (top-to-bottom
as reported)}

\item{short.name}{if TRUE (default) parameters that are not varied
will not be part of the segment type and ID}

\item{id}{if set, the default segment IDs, constructed from numbered
segment types, are replaced by this}

\item{save.matrix}{store the total score matrix \code{S(i,c)} and the
backtracing matrix \code{K(i,c)}; useful in testing stage or for
debugging or illustration of the algorithm
TODO: save.matrix is currently not implemented, since batch function
returns a matrix only}

\item{verb}{level of verbosity, 0: no output, 1: progress messages}

\item{varySettings}{list of settings where each entry can be a vector;
the function will construct a matrix of all possible combinations of
parameter values in this list, call \code{\link{segmentClusters}} for
each, and report a matrix of segments where the segment `type' is
constructed from the varied parameters; see option \code{short.name}.
A varySettings list with all required (default) parameters can be
obtained via function \code{\link{setVarySettings}}.}
}
\description{
A high-level wrapper for multiple runs of segmentation by
\code{\link{segmentClusters}} for multiple clusterings and
multiple segmentation parameters. It additionally allows to
tag adjacent segments to be potentially fused due to similarity
of their clusters.
}
\details{
This is a high-level wrapper for \code{\link{segmentClusters}}
which allows segmentation over multiple clusterings as provided by the
function \code{\link{clusterTimeseries}} and over multiple segmentation
parameters. Each parameter in the list \code{varySettings} can be
a vector and ALL combinations of the passed parameter values will
be used for one run of \code{\link{segmentClusters}}.
}

