% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\docType{package}
\name{segmenTier}
\alias{segmenTier}
\alias{segmenTier-package}
\title{segmenTier : cluster-based segmentation
from a sequential clustering}
\description{
segmenTier : cluster-based segmentation
from a sequential clustering
}
\section{Dependencies}{
 The package strictly depends only on
\code{Rcpp}.
All other dependencies are usually present in a
basic installation (\code{stats}, \code{graphics}, \code{grDevices})).
}

\references{
Machne, Murray & Stadler (2017) <doi:10.1038/s41598-017-12401-8>,
Machne & Murray (2012) <doi:10.1371/journal.pone.0037906>, and
Lehmann et al. (2013) <doi:10.1186/1471-2105-14-133>
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}, Douglas B. Murray,
Peter F. Stadler \email{studla@bioinf.uni-leipzig.de}
}
