% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\docType{package}
\name{segmenTier}
\alias{segmenTier}
\alias{segmenTier-package}
\title{segmenTier : cluster-based segmentation
from a sequential clustering}
\description{
segmenTier : cluster-based segmentation
from a sequential clustering
}
\section{Dependencies}{
 The package strictly depends on \code{package::Rcpp},
and \code{package:parallel} allows to signficantly speed up scoring
function matrix calculations. All other dependencies are
usually present in a basuc installation (\code{graphics}, \code{grDevices})).
}
\author{
Rainer Machne, Peter F. Stadler
}

