% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\docType{package}
\name{segmenTier}
\alias{segmenTier}
\alias{segmenTier-package}
\title{segmenTier : cluster-based segmentation
from a sequential clustering}
\description{
segmenTier : cluster-based segmentation
from a sequential clustering
}
\section{Dependencies}{
 The package strictly depends only on
\code{\link[Rcpp]{Rcpp}}.
All other dependencies are usually present in a
basic installation (\code{stats}, \code{graphics}, \code{grDevices})).
}
\author{
Rainer Machne \email{raim@tbi.univie.ac.at}, Peter F. Stadler \email{studla@bioinf.uni-leipzig.de}
}
\references{
@bibliography segmenTier.bib
}

