% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{sortClusters}
\alias{sortClusters}
\title{Sort clusters by similarity.}
\usage{
sortClusters(cset, sort = TRUE, verb = 0)
}
\arguments{
\item{cset}{a clustering set as returned by \code{\link{clusterTimeseries}}}

\item{sort}{if set to FALSE the clusters will be sorted merely numerically}

\item{verb}{level of verbosity, 0: no output, 1: progress messages}
}
\description{
Takes a clustering set as returned by \code{\link{clusterTimeseries}} and
uses the cluster-cluster similarity matrix \code{Ccc} to sort
clusters by their similarity, starting with the first cluster `1'; the next
cluster is the first cluster (lowest cluster number) with the highest
similarity to cluster `1', and procedding from there. The final sorting is
added as \code{sorting} to \code{cset} and the annotated
\code{cset} is returned. This sorting is subsequently used to select
cluster colors.
}

