% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\docType{data}
\name{tsd}
\alias{tsd}
\title{Transcriptome time-series from budding yeast.}
\description{
Transcriptome time-series data from a region encompassing
four genes and a regulatory upstream non-coding RNA in budding yeast.
The data set is described in more detail in the publication
Machne, Murray & Stadler (2017) <doi:10.1038/s41598-017-12401-8>.
}
