% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\docType{data}
\name{tsd}
\alias{tsd}
\title{transcriptome time series data from a region encompassing
four genes and regulatory upstream non-coding RNA}
\description{
transcriptome time series data from a region encompassing
four genes and regulatory upstream non-coding RNA
}

